Remove new game tab
